Decompose ChatTray.tsx into thin composer + sibling hooks (refs #1446)
Next slice of the multi-PR decomposition tracked by #1446 (after #1459 /
ModernDocumentItem, #1578 / Corpuses, #1586 / ChatMessage, #1580 /
DocumentKnowledgeBase partial).

frontend/src/components/knowledge_base/document/right_tray/ChatTray.tsx had
grown to 1,815 lines. Unlike prior decompositions it owned no inline
styled-components or React sub-components — its weight was concentrated in
stateful handler functions. Three hook extractions land it at 1,160 lines
(-36%):

- useChatStreamHandlers.ts (new): bundles the six chat-state mutators
  (updateMessageApprovalStatus, appendStreamingTokenToChat,
  appendThoughtToMessage, finalizeStreamingResponse, handleCompleteMessage,
  mergeSourcesIntoMessage). useMemo-stabilized return so callers can pass it
  as a single param without rebinding. Param surface kept free of
  documentId/corpusId so CorpusChat.tsx can adopt it in a follow-up.

- useChatAgentMessageHandler.ts (new): wraps the 188-line handleAgentMessage
  switch (10+ message types). Receives pendingApprovalRef (preserves the
  stale-closure escape hatch from the WebSocket subscription) plus the
  stream-handlers bundle.

- useChatSendHandlers.ts (new): bundles sendMessageOverSocket (debounced via
  sendingLockRef), sendApprovalDecision, and sendTextImmediately. Takes
  updateMessageApprovalStatus from the stream hook for optimistic approval UI.

- chatUtils.ts (extended): adds adjustTextareaHeight(textarea, maxHeight) as
  a pure DOM helper.

Behavior preserved verbatim. Public API (ChatTray named export and the
WebSocketSources / MessageData type re-exports) unchanged; importers need no
changes. All 27 ChatTray.ct.tsx tests + 7 ContextMeter.ct.tsx tests pass
without modification.

Render-body splitting is deferred to a follow-up PR — the natural sub-views
share a large overlapping prop surface that has no atom-friendly seam without
more preparatory work.
